requirements.txt
crewai==0.30.11

langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==1.0.3
langsmith==0.1.17

google-api-python-client==2.126.0
arxiv==2.1.0
python-dotenv==1.0.1
streamlit==1.33.0
